+++ rust/gateway/src/main.rs
@@ -114,11 +114,11 @@ async fn run(login: LoginUrl, private_key: StaticSecret) -> Result<Infallible> {
     tunnel
         .set_interface(&init.interface)
         .context("Failed to set interface")?;
-    let mut interface = connlib_shared::interface::TunDeviceManager::default();
-    interface
+    let mut tun_device = connlib_shared::tun_device_manager::TunDeviceManager::new()?;
+    tun_device
         .set_ips(init.interface.ipv4, init.interface.ipv6)
         .await?;
-    interface
+    tun_device
         .set_routes(
             vec![Cidrv4::from(PEERS_IPV4.parse::<Ipv4Network>().unwrap())],
             vec![Cidrv6::from(PEERS_IPV6.parse::<Ipv6Network>().unwrap())],
